update jump double touch

+++ Assets/Script/Netwrok/ConnectServer.cs
@@ -58,7 +58,7 @@ public class ConnectServer : MonoBehaviour {
 			source.Play();		
 			
 			//지정된 시간만큼 음악을 틀어준다
-			int playTime = GlobalScript.getPlayTime();
+			float playTime = GlobalScript.getPlayTime();
 			StartCoroutine(stopMusic(playTime));
 		}
 	}
